refactor(frontend): Move Example and ExampleCategory to EditorModule
