+++ frontend/src/app/editor/four-pane-editor/four-pane-editor.component.ts
@@ -9,8 +9,8 @@ import {MarkdownService} from '../../markdown.service';
 import {Marker} from '../../model/codegen/marker';
 import Request from '../../model/codegen/request';
 import Response from '../../model/codegen/response';
-import Example from '../../model/example';
-import ExampleCategory from '../../model/example-category';
+import {Example} from '../model/example';
+import {ExampleCategory} from '../model/example-category';
 import {PrivacyService} from '../../privacy.service';
 import {ScenarioEditorService} from '../../scenario-editor.service';
 import {ConfigService} from '../config.service';
