+++ frontend/src/app/editor/model/example-category.ts
@@ -1,0 +1,6 @@
+import {Example} from './example';
+
+export class ExampleCategory {
+  name: string;
+  examples: Example[];
+}
