+++ frontend/src/app/editor/model/example.ts
@@ -1,0 +1,7 @@
+import {ExampleCategory} from './example-category';
+
+export class Example {
+  name: string;
+  scenario?: string;
+  category?: ExampleCategory;
+}
