--- frontend/src/app/model/example-category.ts
@@ -1,6 +1,0 @@
-import Example from './example';
-
-export default class ExampleCategory {
-  name: string;
-  examples: Example[];
-}
