--- frontend/src/app/model/example.ts
@@ -1,7 +1,0 @@
-import ExampleCategory from './example-category';
-
-export default class Example {
-  name: string;
-  scenario?: string;
-  category?: ExampleCategory;
-}
